+++ packages/template-retail-react-app/app/pages/checkout-one-click/index.jsx
@@ -18,11 +18,11 @@ import {
 } from '@salesforce/retail-react-app/app/components/shared/ui'
 import useNavigation from '@salesforce/retail-react-app/app/hooks/use-navigation'
 import {useCheckout} from '@salesforce/retail-react-app/app/pages/checkout-container/util/checkout-context'
-import ContactInfo from '@salesforce/retail-react-app/app/pages/checkout-one-click/partials/contact-info'
-import PickupAddress from '@salesforce/retail-react-app/app/pages/checkout-one-click/partials/pickup-address'
-import ShippingAddress from '@salesforce/retail-react-app/app/pages/checkout-one-click/partials/shipping-address'
-import ShippingOptions from '@salesforce/retail-react-app/app/pages/checkout-one-click/partials/shipping-options'
-import Payment from '@salesforce/retail-react-app/app/pages/checkout-one-click/partials/payment'
+import ContactInfo from '@salesforce/retail-react-app/app/pages/checkout-one-click/partials/one-click-contact-info'
+import PickupAddress from '@salesforce/retail-react-app/app/pages/checkout-one-click/partials/one-click-pickup-address'
+import ShippingAddress from '@salesforce/retail-react-app/app/pages/checkout-one-click/partials/one-click-shipping-address'
+import ShippingOptions from '@salesforce/retail-react-app/app/pages/checkout-one-click/partials/one-click-shipping-options'
+import Payment from '@salesforce/retail-react-app/app/pages/checkout-one-click/partials/one-click-payment'
 import OrderSummary from '@salesforce/retail-react-app/app/components/order-summary'
 import {useCurrentBasket} from '@salesforce/retail-react-app/app/hooks/use-current-basket'
 import {useShopperOrdersMutation} from '@salesforce/commerce-sdk-react'
