+++ packages/template-retail-react-app/app/pages/checkout-one-click/partials/one-click-payment.jsx
@@ -17,8 +17,9 @@ import {
     Text,
     Divider
 } from '@salesforce/retail-react-app/app/components/shared/ui'
+import {useForm} from 'react-hook-form'
 import {useToast} from '@salesforce/retail-react-app/app/hooks/use-toast'
-import {useShopperBasketsMutation, useCustomerType} from '@salesforce/commerce-sdk-react'
+import {useShopperBasketsMutation} from '@salesforce/commerce-sdk-react'
 import {useCurrentBasket} from '@salesforce/retail-react-app/app/hooks/use-current-basket'
 import {useCheckout} from '@salesforce/retail-react-app/app/pages/checkout-container/util/checkout-context'
 import {
@@ -33,27 +34,19 @@ import {
 } from '@salesforce/retail-react-app/app/components/toggle-card'
 import PaymentForm from '@salesforce/retail-react-app/app/pages/checkout-one-click/partials/one-click-payment-form'
 import ShippingAddressSelection from '@salesforce/retail-react-app/app/pages/checkout-one-click/partials/one-click-shipping-address-selection'
-import UserRegistration from '@salesforce/retail-react-app/app/pages/checkout-one-click/partials/one-click-user-registration'
 import AddressDisplay from '@salesforce/retail-react-app/app/components/address-display'
 import {PromoCode, usePromoCode} from '@salesforce/retail-react-app/app/components/promo-code'
 import {API_ERROR_MESSAGE} from '@salesforce/retail-react-app/app/constants'
 
-const Payment = ({
-    paymentMethodForm,
-    billingAddressForm,
-    enableUserRegistration,
-    setEnableUserRegistration
-}) => {
+const Payment = () => {
     const {formatMessage} = useIntl()
     const {data: basket} = useCurrentBasket()
-    const {isGuest} = useCustomerType()
     const selectedShippingAddress = basket?.shipments && basket?.shipments[0]?.shippingAddress
     const selectedBillingAddress = basket?.billingAddress
     const appliedPayment = basket?.paymentInstruments && basket?.paymentInstruments[0]
 
     const isPickupOrder = basket?.shipments[0]?.shippingMethod?.c_storePickupEnabled === true
     const [billingSameAsShipping, setBillingSameAsShipping] = useState(!isPickupOrder)
-
     const {mutateAsync: addPaymentInstrumentToBasket} = useShopperBasketsMutation(
         'addPaymentInstrumentToBasket'
     )
@@ -63,20 +56,27 @@ const Payment = ({
     const {mutateAsync: removePaymentInstrumentFromBasket} = useShopperBasketsMutation(
         'removePaymentInstrumentFromBasket'
     )
-
     const showToast = useToast()
-    const showError = (message) => {
+    const showError = () => {
         showToast({
-            title: message || formatMessage(API_ERROR_MESSAGE),
+            title: formatMessage(API_ERROR_MESSAGE),
             status: 'error'
         })
     }
 
-    const {step, STEPS, goToStep} = useCheckout()
+    const {step, STEPS, goToStep, goToNextStep} = useCheckout()
+
+    const billingAddressForm = useForm({
+        mode: 'onChange',
+        shouldUnregister: false,
+        defaultValues: {...selectedBillingAddress}
+    })
 
     // Using destructuring to remove properties from the object...
     // eslint-disable-next-line @typescript-eslint/no-unused-vars
     const {removePromoCode, ...promoCodeProps} = usePromoCode()
+
+    const paymentMethodForm = useForm()
 
     const onPaymentSubmit = async (formValue) => {
         // The form gives us the expiration date as `MM/YY` - so we need to split it into
@@ -99,7 +99,6 @@ const Payment = ({
             body: paymentInstrument
         })
     }
-
     const onBillingSubmit = async () => {
         const isFormValid = await billingAddressForm.trigger()
 
@@ -117,7 +116,6 @@ const Payment = ({
             parameters: {basketId: basket.basketId}
         })
     }
-
     const onPaymentRemoval = async () => {
         try {
             await removePaymentInstrumentFromBasket({
@@ -132,15 +130,16 @@ const Payment = ({
     }
 
     const onSubmit = paymentMethodForm.handleSubmit(async (paymentFormValues) => {
-        try {
-            if (!appliedPayment) {
-                await onPaymentSubmit(paymentFormValues)
-            }
+        if (!appliedPayment) {
+            await onPaymentSubmit(paymentFormValues)
+        }
 
-            // Update billing address
-            await onBillingSubmit()
-        } catch (error) {
-            showError()
+        // If successful `onBillingSubmit` returns the updated basket. If the form was invalid on
+        // submit, `undefined` is returned.
+        const updatedBasket = await onBillingSubmit()
+
+        if (updatedBasket) {
+            goToNextStep()
         }
     })
 
@@ -152,7 +151,6 @@ const Payment = ({
     return (
         <ToggleCard
             id="step-3"
-            data-testid="payment-component"
             title={formatMessage({defaultMessage: 'Payment', id: 'checkout_payment.title.payment'})}
             editing={step === STEPS.PAYMENT}
             isLoading={
@@ -173,7 +171,7 @@ const Payment = ({
 
                 <Stack spacing={6}>
                     {!appliedPayment?.paymentCard ? (
-                        <PaymentForm form={paymentMethodForm} onSubmit={onSubmit} />
+                        <PaymentForm form={paymentMethodForm} onSubmit={onPaymentSubmit} />
                     ) : (
                         <Stack spacing={3}>
                             <Heading as="h3" fontSize="md">
@@ -209,7 +207,7 @@ const Payment = ({
                             />
                         </Heading>
 
-                        {!isPickupOrder && selectedShippingAddress && (
+                        {!isPickupOrder && (
                             <Checkbox
                                 name="billingSameAsShipping"
                                 isChecked={billingSameAsShipping}
@@ -240,12 +238,17 @@ const Payment = ({
                             isBillingAddress
                         />
                     )}
-                    {isGuest && (
-                        <UserRegistration
-                            enableUserRegistration={enableUserRegistration}
-                            setEnableUserRegistration={setEnableUserRegistration}
-                        />
-                    )}
+
+                    <Box pt={3}>
+                        <Container variant="form">
+                            <Button w="full" onClick={onSubmit}>
+                                <FormattedMessage
+                                    defaultMessage="Review Order"
+                                    id="checkout_payment.button.review_order"
+                                />
+                            </Button>
+                        </Container>
+                    </Box>
                 </Stack>
             </ToggleCardEdit>
 
@@ -276,22 +279,10 @@ const Payment = ({
                             <AddressDisplay address={selectedBillingAddress} />
                         </Stack>
                     )}
-
-                    <UserRegistration
-                        enableUserRegistration={enableUserRegistration}
-                        setEnableUserRegistration={setEnableUserRegistration}
-                    />
                 </Stack>
             </ToggleCardSummary>
         </ToggleCard>
     )
-}
-
-Payment.propTypes = {
-    /** Whether user registration is enabled */
-    enableUserRegistration: PropTypes.bool,
-    /** Callback to set user registration state */
-    setEnableUserRegistration: PropTypes.func
 }
 
 const PaymentCardSummary = ({payment}) => {
@@ -312,10 +303,5 @@ const PaymentCardSummary = ({payment}) => {
 }
 
 PaymentCardSummary.propTypes = {payment: PropTypes.object}
-
-Payment.propTypes = {
-    paymentMethodForm: PropTypes.object.isRequired,
-    billingAddressForm: PropTypes.object.isRequired
-}
 
 export default Payment
