+++ packages/template-retail-react-app/app/pages/checkout-one-click/partials/one-click-shipping-address.jsx
@@ -4,7 +4,7 @@
  * SPDX-License-Identifier: BSD-3-Clause
  * For full license text, see the LICENSE file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
-import React, {useState, useEffect} from 'react'
+import React, {useState} from 'react'
 import {nanoid} from 'nanoid'
 import {defineMessage, useIntl} from 'react-intl'
 import {useCheckout} from '@salesforce/retail-react-app/app/pages/checkout-container/util/checkout-context'
@@ -34,7 +34,6 @@ const shippingAddressAriaLabel = defineMessage({
 export default function ShippingAddress() {
     const {formatMessage} = useIntl()
     const [isLoading, setIsLoading] = useState()
-    const [hasAutoSelected, setHasAutoSelected] = useState(false)
     const {data: customer} = useCurrentCustomer()
     const {data: basket} = useCurrentBasket()
     const selectedShippingAddress = basket?.shipments && basket?.shipments[0]?.shippingAddress
@@ -48,9 +47,24 @@ export default function ShippingAddress() {
 
     const submitAndContinue = async (address) => {
         setIsLoading(true)
-        try {
-            const {
-                addressId,
+        const {
+            addressId,
+            address1,
+            city,
+            countryCode,
+            firstName,
+            lastName,
+            phone,
+            postalCode,
+            stateCode
+        } = address
+        await updateShippingAddressForShipment.mutateAsync({
+            parameters: {
+                basketId: basket.basketId,
+                shipmentId: 'me',
+                useAsBilling: false
+            },
+            body: {
                 address1,
                 city,
                 countryCode,
@@ -59,100 +73,40 @@ export default function ShippingAddress() {
                 phone,
                 postalCode,
                 stateCode
-            } = address
-            await updateShippingAddressForShipment.mutateAsync({
+            }
+        })
+
+        if (customer.isRegistered && !addressId) {
+            const body = {
+                address1,
+                city,
+                countryCode,
+                firstName,
+                lastName,
+                phone,
+                postalCode,
+                stateCode,
+                addressId: nanoid()
+            }
+            await createCustomerAddress.mutateAsync({
+                body,
+                parameters: {customerId: customer.customerId}
+            })
+        }
+
+        if (customer.isRegistered && addressId) {
+            await updateCustomerAddress.mutateAsync({
+                body: address,
                 parameters: {
-                    basketId: basket.basketId,
-                    shipmentId: 'me',
-                    useAsBilling: false
-                },
-                body: {
-                    address1,
-                    city,
-                    countryCode,
-                    firstName,
-                    lastName,
-                    phone,
-                    postalCode,
-                    stateCode
+                    customerId: customer.customerId,
+                    addressName: addressId
                 }
             })
-
-            if (customer.isRegistered && !addressId) {
-                const body = {
-                    address1,
-                    city,
-                    countryCode,
-                    firstName,
-                    lastName,
-                    phone,
-                    postalCode,
-                    stateCode,
-                    addressId: nanoid()
-                }
-                await createCustomerAddress.mutateAsync({
-                    body,
-                    parameters: {customerId: customer.customerId}
-                })
-            }
-
-            if (customer.isRegistered && addressId) {
-                await updateCustomerAddress.mutateAsync({
-                    body: address,
-                    parameters: {
-                        customerId: customer.customerId,
-                        addressName: addressId
-                    }
-                })
-            }
-
-            goToNextStep()
-        } catch (error) {
-            console.error('Error submitting shipping address:', error)
-        } finally {
-            setIsLoading(false)
         }
+
+        goToNextStep()
+        setIsLoading(false)
     }
-
-    // Auto-select and apply preferred shipping address when component is on this step
-    useEffect(() => {
-        const autoSelectPreferredAddress = async () => {
-            // Only auto-select when on this step and haven't already auto-selected
-            if (step !== STEPS.SHIPPING_ADDRESS || hasAutoSelected || isLoading) {
-                return
-            }
-
-            // Only proceed if customer is registered and has addresses
-            if (!customer?.isRegistered || !customer?.addresses?.length) {
-                return
-            }
-
-            // Skip to next step if basket already has a shipping address
-            if (selectedShippingAddress?.address1) {
-                setHasAutoSelected(true) // Prevent further attempts
-                goToNextStep()
-                return
-            }
-
-            // Find the preferred address
-            const preferredAddress = customer.addresses.find((addr) => addr.preferred === true)
-
-            //Auto-selecting preferred shipping address
-            if (preferredAddress) {
-                setHasAutoSelected(true)
-
-                try {
-                    // Apply the preferred address and continue to next step
-                    await submitAndContinue(preferredAddress)
-                } catch (error) {
-                    // Reset on error so user can manually select
-                    setHasAutoSelected(false)
-                }
-            }
-        }
-
-        autoSelectPreferredAddress()
-    }, [step, customer, selectedShippingAddress, hasAutoSelected, isLoading])
 
     return (
         <ToggleCard
