+++ packages/template-retail-react-app/app/pages/checkout-one-click/partials/one-click-shipping-options.jsx
@@ -4,7 +4,7 @@
  * SPDX-License-Identifier: BSD-3-Clause
  * For full license text, see the LICENSE file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
-import React, {useEffect, useState, useMemo} from 'react'
+import React, {useEffect} from 'react'
 import {FormattedMessage, FormattedNumber, useIntl} from 'react-intl'
 import {
     Box,
@@ -29,18 +29,14 @@ import {
     useShopperBasketsMutation
 } from '@salesforce/commerce-sdk-react'
 import {useCurrentBasket} from '@salesforce/retail-react-app/app/hooks/use-current-basket'
-import {useCurrentCustomer} from '@salesforce/retail-react-app/app/hooks/use-current-customer'
 import {useCurrency} from '@salesforce/retail-react-app/app/hooks'
 
 export default function ShippingOptions() {
     const {formatMessage} = useIntl()
     const {step, STEPS, goToStep, goToNextStep} = useCheckout()
     const {data: basket} = useCurrentBasket()
-    const {data: customer} = useCurrentCustomer()
     const {currency} = useCurrency()
     const updateShippingMethod = useShopperBasketsMutation('updateShippingMethodForShipment')
-    const [hasAutoSelected, setHasAutoSelected] = useState(false)
-    const [isLoading, setIsLoading] = useState(false)
     const {data: shippingMethods} = useShippingMethodsForShipment(
         {
             parameters: {
@@ -56,24 +52,6 @@ export default function ShippingOptions() {
     const selectedShippingMethod = basket?.shipments?.[0]?.shippingMethod
     const selectedShippingAddress = basket?.shipments?.[0]?.shippingAddress
 
-    // Calculate if we should show loading state immediately for auto-selection
-    const shouldShowInitialLoading = useMemo(() => {
-        return (
-            step === STEPS.SHIPPING_OPTIONS &&
-            !hasAutoSelected &&
-            customer?.isRegistered &&
-            !selectedShippingMethod?.id &&
-            shippingMethods?.applicableShippingMethods?.length &&
-            shippingMethods.defaultShippingMethodId &&
-            shippingMethods.applicableShippingMethods.find(
-                (method) => method.id === shippingMethods.defaultShippingMethodId
-            )
-        )
-    }, [step, hasAutoSelected, customer, selectedShippingMethod, shippingMethods])
-
-    // Use calculated loading state or manual loading state
-    const effectiveIsLoading = isLoading || shouldShowInitialLoading
-
     const form = useForm({
         shouldUnregister: false,
         defaultValues: {
@@ -87,71 +65,10 @@ export default function ShippingOptions() {
         if (!selectedShippingMethod && !methodId && defaultMethodId) {
             form.reset({shippingMethodId: defaultMethodId})
         }
-
         if (selectedShippingMethod && methodId !== selectedShippingMethod.id) {
             form.reset({shippingMethodId: selectedShippingMethod.id})
         }
     }, [selectedShippingMethod, shippingMethods])
-
-    // Auto-select default shipping method and proceed for authenticated users
-    useEffect(() => {
-        const autoSelectDefaultShippingMethod = async () => {
-            // Only auto-select when on this step and haven't already auto-selected
-            if (step !== STEPS.SHIPPING_OPTIONS || hasAutoSelected || isLoading) {
-                return
-            }
-
-            // Skip if basket already has a shipping method
-            if (selectedShippingMethod?.id) {
-                setHasAutoSelected(true)
-                goToNextStep()
-                return
-            }
-
-            // Only proceed for authenticated users
-            if (!customer?.isRegistered) {
-                return
-            }
-
-            // Wait for shipping methods to load
-            if (!shippingMethods?.applicableShippingMethods?.length) {
-                return
-            }
-
-            const defaultMethodId = shippingMethods.defaultShippingMethodId
-            const defaultMethod = shippingMethods.applicableShippingMethods.find(
-                (method) => method.id === defaultMethodId
-            )
-
-            if (defaultMethod) {
-                //Auto-selecting default shipping method
-                setHasAutoSelected(true)
-                setIsLoading(true) // Show loading state immediately
-
-                try {
-                    // Apply the default shipping method and continue to next step
-                    await updateShippingMethod.mutateAsync({
-                        parameters: {
-                            basketId: basket.basketId,
-                            shipmentId: 'me'
-                        },
-                        body: {
-                            id: defaultMethodId
-                        }
-                    })
-                    //Default shipping method auto-applied successfully
-                    setIsLoading(false) // Clear loading state before navigation
-                    goToNextStep()
-                } catch (error) {
-                    // Reset on error so user can manually select
-                    setHasAutoSelected(false)
-                    setIsLoading(false) // Hide loading state on error
-                }
-            }
-        }
-
-        autoSelectDefaultShippingMethod()
-    }, [step, selectedShippingMethod, customer, shippingMethods, hasAutoSelected, basket?.basketId])
 
     const submitForm = async ({shippingMethodId}) => {
         await updateShippingMethod.mutateAsync({
@@ -207,10 +124,8 @@ export default function ShippingOptions() {
                 id: 'shipping_options.title.shipping_gift_options'
             })}
             editing={step === STEPS.SHIPPING_OPTIONS}
-            isLoading={form.formState.isSubmitting || effectiveIsLoading}
-            disabled={
-                selectedShippingMethod == null || !selectedShippingAddress || effectiveIsLoading
-            }
+            isLoading={form.formState.isSubmitting}
+            disabled={selectedShippingMethod == null || !selectedShippingAddress}
             onEdit={() => goToStep(STEPS.SHIPPING_OPTIONS)}
             editLabel={formatMessage({
                 defaultMessage: 'Edit Shipping Options',
@@ -299,7 +214,7 @@ export default function ShippingOptions() {
                 </form>
             </ToggleCardEdit>
 
-            {!effectiveIsLoading && selectedShippingMethod && selectedShippingAddress && (
+            {selectedShippingMethod && selectedShippingAddress && (
                 <ToggleCardSummary>
                     <Flex justify="space-between" w="full">
                         <Text>{selectedShippingMethod.name}</Text>
