--- packages/template-retail-react-app/app/pages/checkout-one-click/partials/payment.jsx
@@ -1,307 +1,0 @@
-/*
- * Copyright (c) 2021, salesforce.com, inc.
- * All rights reserved.
- * SPDX-License-Identifier: BSD-3-Clause
- * For full license text, see the LICENSE file in the repo root or https://opensource.org/licenses/BSD-3-Clause
- */
-import React, {useState} from 'react'
-import PropTypes from 'prop-types'
-import {defineMessage, FormattedMessage, useIntl} from 'react-intl'
-import {
-    Box,
-    Button,
-    Checkbox,
-    Container,
-    Heading,
-    Stack,
-    Text,
-    Divider
-} from '@salesforce/retail-react-app/app/components/shared/ui'
-import {useForm} from 'react-hook-form'
-import {useToast} from '@salesforce/retail-react-app/app/hooks/use-toast'
-import {useShopperBasketsMutation} from '@salesforce/commerce-sdk-react'
-import {useCurrentBasket} from '@salesforce/retail-react-app/app/hooks/use-current-basket'
-import {useCheckout} from '@salesforce/retail-react-app/app/pages/checkout-container/util/checkout-context'
-import {
-    getPaymentInstrumentCardType,
-    getMaskCreditCardNumber,
-    getCreditCardIcon
-} from '@salesforce/retail-react-app/app/utils/cc-utils'
-import {
-    ToggleCard,
-    ToggleCardEdit,
-    ToggleCardSummary
-} from '@salesforce/retail-react-app/app/components/toggle-card'
-import PaymentForm from '@salesforce/retail-react-app/app/pages/checkout-one-click/partials/payment-form'
-import ShippingAddressSelection from '@salesforce/retail-react-app/app/pages/checkout-one-click/partials/shipping-address-selection'
-import AddressDisplay from '@salesforce/retail-react-app/app/components/address-display'
-import {PromoCode, usePromoCode} from '@salesforce/retail-react-app/app/components/promo-code'
-import {API_ERROR_MESSAGE} from '@salesforce/retail-react-app/app/constants'
-
-const Payment = () => {
-    const {formatMessage} = useIntl()
-    const {data: basket} = useCurrentBasket()
-    const selectedShippingAddress = basket?.shipments && basket?.shipments[0]?.shippingAddress
-    const selectedBillingAddress = basket?.billingAddress
-    const appliedPayment = basket?.paymentInstruments && basket?.paymentInstruments[0]
-
-    const isPickupOrder = basket?.shipments[0]?.shippingMethod?.c_storePickupEnabled === true
-    const [billingSameAsShipping, setBillingSameAsShipping] = useState(!isPickupOrder)
-    const {mutateAsync: addPaymentInstrumentToBasket} = useShopperBasketsMutation(
-        'addPaymentInstrumentToBasket'
-    )
-    const {mutateAsync: updateBillingAddressForBasket} = useShopperBasketsMutation(
-        'updateBillingAddressForBasket'
-    )
-    const {mutateAsync: removePaymentInstrumentFromBasket} = useShopperBasketsMutation(
-        'removePaymentInstrumentFromBasket'
-    )
-    const showToast = useToast()
-    const showError = () => {
-        showToast({
-            title: formatMessage(API_ERROR_MESSAGE),
-            status: 'error'
-        })
-    }
-
-    const {step, STEPS, goToStep, goToNextStep} = useCheckout()
-
-    const billingAddressForm = useForm({
-        mode: 'onChange',
-        shouldUnregister: false,
-        defaultValues: {...selectedBillingAddress}
-    })
-
-    // Using destructuring to remove properties from the object...
-    // eslint-disable-next-line @typescript-eslint/no-unused-vars
-    const {removePromoCode, ...promoCodeProps} = usePromoCode()
-
-    const paymentMethodForm = useForm()
-
-    const onPaymentSubmit = async (formValue) => {
-        // The form gives us the expiration date as `MM/YY` - so we need to split it into
-        // month and year to submit them as individual fields.
-        const [expirationMonth, expirationYear] = formValue.expiry.split('/')
-
-        const paymentInstrument = {
-            paymentMethodId: 'CREDIT_CARD',
-            paymentCard: {
-                holder: formValue.holder,
-                maskedNumber: getMaskCreditCardNumber(formValue.number),
-                cardType: getPaymentInstrumentCardType(formValue.cardType),
-                expirationMonth: parseInt(expirationMonth),
-                expirationYear: parseInt(`20${expirationYear}`)
-            }
-        }
-
-        return addPaymentInstrumentToBasket({
-            parameters: {basketId: basket?.basketId},
-            body: paymentInstrument
-        })
-    }
-    const onBillingSubmit = async () => {
-        const isFormValid = await billingAddressForm.trigger()
-
-        if (!isFormValid) {
-            return
-        }
-        const billingAddress = billingSameAsShipping
-            ? selectedShippingAddress
-            : billingAddressForm.getValues()
-        // Using destructuring to remove properties from the object...
-        // eslint-disable-next-line @typescript-eslint/no-unused-vars
-        const {addressId, creationDate, lastModified, preferred, ...address} = billingAddress
-        return await updateBillingAddressForBasket({
-            body: address,
-            parameters: {basketId: basket.basketId}
-        })
-    }
-    const onPaymentRemoval = async () => {
-        try {
-            await removePaymentInstrumentFromBasket({
-                parameters: {
-                    basketId: basket.basketId,
-                    paymentInstrumentId: appliedPayment.paymentInstrumentId
-                }
-            })
-        } catch (e) {
-            showError()
-        }
-    }
-
-    const onSubmit = paymentMethodForm.handleSubmit(async (paymentFormValues) => {
-        if (!appliedPayment) {
-            await onPaymentSubmit(paymentFormValues)
-        }
-
-        // If successful `onBillingSubmit` returns the updated basket. If the form was invalid on
-        // submit, `undefined` is returned.
-        const updatedBasket = await onBillingSubmit()
-
-        if (updatedBasket) {
-            goToNextStep()
-        }
-    })
-
-    const billingAddressAriaLabel = defineMessage({
-        defaultMessage: 'Billing Address Form',
-        id: 'checkout_payment.label.billing_address_form'
-    })
-
-    return (
-        <ToggleCard
-            id="step-3"
-            title={formatMessage({defaultMessage: 'Payment', id: 'checkout_payment.title.payment'})}
-            editing={step === STEPS.PAYMENT}
-            isLoading={
-                paymentMethodForm.formState.isSubmitting ||
-                billingAddressForm.formState.isSubmitting
-            }
-            disabled={appliedPayment == null}
-            onEdit={() => goToStep(STEPS.PAYMENT)}
-            editLabel={formatMessage({
-                defaultMessage: 'Edit Payment Info',
-                id: 'toggle_card.action.editPaymentInfo'
-            })}
-        >
-            <ToggleCardEdit>
-                <Box mt={-2} mb={4}>
-                    <PromoCode {...promoCodeProps} itemProps={{border: 'none'}} />
-                </Box>
-
-                <Stack spacing={6}>
-                    {!appliedPayment?.paymentCard ? (
-                        <PaymentForm form={paymentMethodForm} onSubmit={onPaymentSubmit} />
-                    ) : (
-                        <Stack spacing={3}>
-                            <Heading as="h3" fontSize="md">
-                                <FormattedMessage
-                                    defaultMessage="Credit Card"
-                                    id="checkout_payment.heading.credit_card"
-                                />
-                            </Heading>
-                            <Stack direction="row" spacing={4}>
-                                <PaymentCardSummary payment={appliedPayment} />
-                                <Button
-                                    variant="link"
-                                    size="sm"
-                                    colorScheme="red"
-                                    onClick={onPaymentRemoval}
-                                >
-                                    <FormattedMessage
-                                        defaultMessage="Remove"
-                                        id="checkout_payment.action.remove"
-                                    />
-                                </Button>
-                            </Stack>
-                        </Stack>
-                    )}
-
-                    <Divider borderColor="gray.100" />
-
-                    <Stack spacing={2}>
-                        <Heading as="h3" fontSize="md">
-                            <FormattedMessage
-                                defaultMessage="Billing Address"
-                                id="checkout_payment.heading.billing_address"
-                            />
-                        </Heading>
-
-                        {!isPickupOrder && (
-                            <Checkbox
-                                name="billingSameAsShipping"
-                                isChecked={billingSameAsShipping}
-                                onChange={(e) => setBillingSameAsShipping(e.target.checked)}
-                            >
-                                <Text fontSize="sm" color="gray.700">
-                                    <FormattedMessage
-                                        defaultMessage="Same as shipping address"
-                                        id="checkout_payment.label.same_as_shipping"
-                                    />
-                                </Text>
-                            </Checkbox>
-                        )}
-
-                        {billingSameAsShipping && selectedShippingAddress && (
-                            <Box pl={7}>
-                                <AddressDisplay address={selectedShippingAddress} />
-                            </Box>
-                        )}
-                    </Stack>
-
-                    {!billingSameAsShipping && (
-                        <ShippingAddressSelection
-                            form={billingAddressForm}
-                            selectedAddress={selectedBillingAddress}
-                            formTitleAriaLabel={billingAddressAriaLabel}
-                            hideSubmitButton
-                            isBillingAddress
-                        />
-                    )}
-
-                    <Box pt={3}>
-                        <Container variant="form">
-                            <Button w="full" onClick={onSubmit}>
-                                <FormattedMessage
-                                    defaultMessage="Review Order"
-                                    id="checkout_payment.button.review_order"
-                                />
-                            </Button>
-                        </Container>
-                    </Box>
-                </Stack>
-            </ToggleCardEdit>
-
-            <ToggleCardSummary>
-                <Stack spacing={6}>
-                    {appliedPayment && (
-                        <Stack spacing={3}>
-                            <Heading as="h3" fontSize="md">
-                                <FormattedMessage
-                                    defaultMessage="Credit Card"
-                                    id="checkout_payment.heading.credit_card"
-                                />
-                            </Heading>
-                            <PaymentCardSummary payment={appliedPayment} />
-                        </Stack>
-                    )}
-
-                    <Divider borderColor="gray.100" />
-
-                    {selectedBillingAddress && (
-                        <Stack spacing={2}>
-                            <Heading as="h3" fontSize="md">
-                                <FormattedMessage
-                                    defaultMessage="Billing Address"
-                                    id="checkout_payment.heading.billing_address"
-                                />
-                            </Heading>
-                            <AddressDisplay address={selectedBillingAddress} />
-                        </Stack>
-                    )}
-                </Stack>
-            </ToggleCardSummary>
-        </ToggleCard>
-    )
-}
-
-const PaymentCardSummary = ({payment}) => {
-    const CardIcon = getCreditCardIcon(payment?.paymentCard?.cardType)
-    return (
-        <Stack direction="row" alignItems="center" spacing={3}>
-            {CardIcon && <CardIcon layerStyle="ccIcon" />}
-
-            <Stack direction="row">
-                <Text>{payment.paymentCard.cardType}</Text>
-                <Text>&bull;&bull;&bull;&bull; {payment.paymentCard.numberLastDigits}</Text>
-                <Text>
-                    {payment.paymentCard.expirationMonth}/{payment.paymentCard.expirationYear}
-                </Text>
-            </Stack>
-        </Stack>
-    )
-}
-
-PaymentCardSummary.propTypes = {payment: PropTypes.object}
-
-export default Payment
